[Refactor]: Move fmtutil from cmd to pkg/formatter
