+++ cmd/nerdctl/apparmor_ls_linux.go
@@ -24,6 +24,7 @@ import (
 	"text/template"
 
 	"github.com/containerd/nerdctl/pkg/apparmorutil"
+	"github.com/containerd/nerdctl/pkg/formatter"
 	"github.com/spf13/cobra"
 )
 
@@ -70,7 +71,7 @@ func apparmorLsAction(cmd *cobra.Command, args []string) error {
 			return errors.New("format and quiet must not be specified together")
 		}
 		var err error
-		tmpl, err = parseTemplate(format)
+		tmpl, err = formatter.ParseTemplate(format)
 		if err != nil {
 			return err
 		}
@@ -96,7 +97,7 @@ func apparmorLsAction(cmd *cobra.Command, args []string) error {
 			fmt.Fprintf(w, "%s\t%s\n", f.Name, f.Mode)
 		}
 	}
-	if f, ok := w.(Flusher); ok {
+	if f, ok := w.(formatter.Flusher); ok {
 		return f.Flush()
 	}
 	return nil
