+++ cmd/nerdctl/events.go
@@ -26,6 +26,7 @@ import (
 
 	"github.com/containerd/containerd/events"
 	"github.com/containerd/containerd/log"
+	"github.com/containerd/nerdctl/pkg/formatter"
 	"github.com/containerd/typeurl"
 
 	"github.com/spf13/cobra"
@@ -81,7 +82,7 @@ func eventsAction(cmd *cobra.Command, args []string) error {
 	case "raw", "table", "wide":
 		return errors.New("unsupported format: \"raw\", \"table\", and \"wide\"")
 	default:
-		tmpl, err = parseTemplate(format)
+		tmpl, err = formatter.ParseTemplate(format)
 		if err != nil {
 			return err
 		}
