+++ cmd/nerdctl/history.go
@@ -195,7 +195,7 @@ func printHistory(cmd *cobra.Command, historys []historyPrintable) error {
 			return errors.New("format and quiet must not be specified together")
 		}
 		var err error
-		tmpl, err = parseTemplate(format)
+		tmpl, err = formatter.ParseTemplate(format)
 		if err != nil {
 			return err
 		}
@@ -214,7 +214,7 @@ func printHistory(cmd *cobra.Command, historys []historyPrintable) error {
 		}
 	}
 
-	if f, ok := w.(Flusher); ok {
+	if f, ok := w.(formatter.Flusher); ok {
 		return f.Flush()
 	}
 	return nil
