+++ cmd/nerdctl/volume_ls.go
@@ -26,6 +26,7 @@ import (
 	"text/template"
 
 	"github.com/containerd/containerd/pkg/progress"
+	"github.com/containerd/nerdctl/pkg/formatter"
 	"github.com/containerd/nerdctl/pkg/inspecttypes/native"
 	"github.com/sirupsen/logrus"
 
@@ -116,7 +117,7 @@ func volumeLsAction(cmd *cobra.Command, args []string) error {
 			return errors.New("format and quiet must not be specified together")
 		}
 		var err error
-		tmpl, err = parseTemplate(format)
+		tmpl, err = formatter.ParseTemplate(format)
 		if err != nil {
 			return err
 		}
@@ -139,7 +140,7 @@ func volumeLsAction(cmd *cobra.Command, args []string) error {
 			Scope:      "local",
 		}
 		if v.Labels != nil {
-			p.Labels = formatLabels(*v.Labels)
+			p.Labels = formatter.FormatLabels(*v.Labels)
 		}
 		if volumeSize {
 			p.Size = progress.Bytes(v.Size).String()
@@ -160,7 +161,7 @@ func volumeLsAction(cmd *cobra.Command, args []string) error {
 			fmt.Fprintf(w, "%s\t%s\n", p.Name, p.Mountpoint)
 		}
 	}
-	if f, ok := w.(Flusher); ok {
+	if f, ok := w.(formatter.Flusher); ok {
 		return f.Flush()
 	}
 	return nil
